+++ lib/src/functions.dart
@@ -267,8 +267,7 @@ abstract class DefaultFunction extends MathFunction {
   }
 
   @override
-  String toString() =>
-      args.length == 1 ? '$name(${args[0]})' : '$name(${args[0]},${args[1]})';
+  String toString() => '$name(${args.join(',')})';
 }
 
 /// The exponential function.
@@ -285,7 +284,7 @@ class Exponential extends DefaultFunction {
   ///     x = Variable('x');
   ///     exp = Exponential(x);
   ///     exp = Exponential(x + four);
-  Exponential(Expression exp) : super._unary('exp', exp);
+  Exponential(Expression exp) : super._unary('e', exp);
 
   /// The exponent of this exponential function.
   Expression get exp => getParam(0);
@@ -382,9 +381,6 @@ class Log extends DefaultFunction {
     throw UnimplementedError('Can not evaluate $name on $type yet.');
   }
 
-  @override
-  String toString() => 'log_$base($arg)';
-
   /// Returns the natural from of this logarithm.
   /// E.g. log_10(2) = ln(2) / ln(10)
   ///
@@ -453,10 +449,10 @@ class Root extends DefaultFunction {
   /// For example, to create the 5th root of x:
   ///
   ///     root = Root(5, Variable('x'));
-  Root(this.n, Expression arg) : super._unary('root', arg);
+  Root(this.n, Expression arg) : super._unary('nrt', arg);
 
   /// Creates the n-th root of arg where n is a [Number] literal.
-  Root.fromExpr(Number n, Expression arg) : super._unary('root', arg) {
+  Root.fromExpr(Number n, Expression arg) : super._unary('nrt', arg) {
     this.n = n.getConstantValue().toInt();
   }
 
@@ -464,7 +460,7 @@ class Root extends DefaultFunction {
   ///
   /// For example, to create the square root of x:
   ///
-  ///     sqrt = Root.sqrt(new Variable('x'));
+  ///     sqrt = Root.sqrt(Variable('x'));
   ///
   /// __Note__:
   /// For better simplification and display, use the [Sqrt] class.
@@ -486,7 +482,7 @@ class Root extends DefaultFunction {
       this.asPower().evaluate(type, context);
 
   @override
-  String toString() => 'nrt_$n($arg)';
+  String toString() => 'nrt($n,$arg)';
 
   /// Returns the power form of this root.
   /// E.g. root_5(x) = x^(1/5)
@@ -502,7 +498,7 @@ class Sqrt extends Root {
   ///
   /// For example, to create the square root of x:
   ///
-  ///     sqrt = Sqrt(new Variable('x'));
+  ///     sqrt = Sqrt(Variable('x'));
   Sqrt(Expression arg) : super.sqrt(arg);
 
   /// Possible simplifications: