+++ test/parser_test_set.dart
@@ -12,7 +12,9 @@ class ParserTests extends TestSet {
         'Lexer Tokenize (Infix + Postfix)': lexerTokenTest,
         'Lexer Tokenize Invalid': lexerTokenTestInvalid,
         'Parser Expression Creation': parserExpressionTest,
-        'Parser Expression Creation Invalid': parserExpressionTestInvalid
+        'Parser Expression Creation Invalid': parserExpressionTestInvalid,
+        'Parser Expression Creation from toString()':
+            parserExpressionTest_ParseFromToString,
       };
 
   @override
@@ -104,34 +106,6 @@ class ParserTests extends TestSet {
       Token('1', TokenType.VAL),
       Token('/', TokenType.DIV)
     ]);
-
-    // Unary Minus
-    // underscoe syntax
-    inputStrings.add('_1');
-    tokenStreams
-        .add([Token('_', TokenType.UNMINUS), Token('1', TokenType.VAL)]);
-    rpnTokenStreams
-        .add([Token('1', TokenType.VAL), Token('_', TokenType.UNMINUS)]);
-
-    inputStrings.add('(_1)');
-    tokenStreams.add([
-      Token('(', TokenType.LBRACE),
-      Token('_', TokenType.UNMINUS),
-      Token('1', TokenType.VAL),
-      Token(')', TokenType.RBRACE)
-    ]);
-    rpnTokenStreams
-        .add([Token('1', TokenType.VAL), Token('_', TokenType.UNMINUS)]);
-
-    inputStrings.add('_(1)');
-    tokenStreams.add([
-      Token('_', TokenType.UNMINUS),
-      Token('(', TokenType.LBRACE),
-      Token('1', TokenType.VAL),
-      Token(')', TokenType.RBRACE)
-    ]);
-    rpnTokenStreams
-        .add([Token('1', TokenType.VAL), Token('_', TokenType.UNMINUS)]);
 
     // standard syntax
     inputStrings.add('-1');
@@ -251,26 +225,26 @@ class ParserTests extends TestSet {
         .add([Token('10', TokenType.VAL), Token('tan', TokenType.TAN)]);
 
     // Arccos
-    inputStrings.add('arccos(10)');
+    inputStrings.add('arccos(1)');
     tokenStreams.add([
       Token('arccos', TokenType.ACOS),
       Token('(', TokenType.LBRACE),
-      Token('10', TokenType.VAL),
+      Token('1', TokenType.VAL),
       Token(')', TokenType.RBRACE)
     ]);
     rpnTokenStreams
-        .add([Token('10', TokenType.VAL), Token('arccos', TokenType.ACOS)]);
+        .add([Token('1', TokenType.VAL), Token('arccos', TokenType.ACOS)]);
 
     // Arcsin
-    inputStrings.add('arcsin(10)');
+    inputStrings.add('arcsin(1)');
     tokenStreams.add([
       Token('arcsin', TokenType.ASIN),
       Token('(', TokenType.LBRACE),
-      Token('10', TokenType.VAL),
+      Token('1', TokenType.VAL),
       Token(')', TokenType.RBRACE)
     ]);
     rpnTokenStreams
-        .add([Token('10', TokenType.VAL), Token('arcsin', TokenType.ASIN)]);
+        .add([Token('1', TokenType.VAL), Token('arcsin', TokenType.ASIN)]);
 
     // Arctan
     inputStrings.add('arctan(10)');
@@ -428,11 +402,11 @@ class ParserTests extends TestSet {
 
   void lexerTokenTestInvalid() {
     Map<String, Matcher> invalidCases = {
-      '(': throwsStateError,
-      ')': throwsStateError,
-      '1+1)': throwsStateError,
-      '(1+1': throwsStateError,
-      'log(1,': throwsStateError,
+      '(': throwsFormatException,
+      ')': throwsFormatException,
+      '1+1)': throwsFormatException,
+      '(1+1': throwsFormatException,
+      'log(1,': throwsFormatException,
     };
 
     for (String expr in invalidCases.keys) {
@@ -445,22 +419,49 @@ class ParserTests extends TestSet {
       Expression exp = pars.parse(inputString);
       // TODO Don't just test for no exceptions,
       // also test for expression content.
+      expect(exp, isNotNull);
     }
   }
 
   void parserExpressionTestInvalid() {
     Map<String, Matcher> invalidCases = {
-      '': throwsArgumentError,
-      '(': throwsStateError,
-      ')': throwsStateError,
-      '1+1)': throwsStateError,
-      '(1+1': throwsStateError,
+      '': throwsFormatException,
+      '(': throwsFormatException,
+      ')': throwsFormatException,
+      '1+1)': throwsFormatException,
+      '(1+1': throwsFormatException,
       'log(,1)': throwsRangeError,
       'log(1,)': throwsRangeError,
     };
 
     for (String expr in invalidCases.keys) {
       expect(() => pars.parse(expr), invalidCases[expr]);
+    }
+  }
+
+  void parserExpressionTest_ParseFromToString() {
+    ContextModel context = ContextModel()
+      ..bindVariableName('x', Number(math.pi));
+
+    for (String inputString in inputStrings) {
+      /// Expression doesn't implement equal, so as an approximation
+      /// we're testing whether the expression re-parses and evaluates
+      /// to the same value.
+      Expression exp = pars.parse(inputString);
+
+      try {
+        Expression exp2 = pars.parse(exp.toString());
+
+        double r1 = exp.evaluate(EvaluationType.REAL, context);
+        double r2 = exp2.evaluate(EvaluationType.REAL, context);
+        expect(r2, r1, reason: 'Expected ${r2} for ${exp} (${exp2})');
+      } on FormatException catch (fe) {
+        expect(fe, isNot(isFormatException),
+            reason: 'Expected no exception for ${inputString} (${exp})');
+      } on RangeError catch (re) {
+        expect(re, isNot(isRangeError),
+            reason: 'Expected no exception for ${inputString} (${exp})');
+      }
     }
   }
 }